\docType{data}
\name{condprobnum}
\alias{condprobnum}
\title{Conditional probability values used by InterVA-4}
\format{A data frame with 245 observations on 60 variables. Each observation
is the conditional probability.}
\description{
This is a conditional probability matrix used by InterVA-4.2.
There are 60 causes and 245 symptoms.  
}
\examples{
data(condprobnum)
}
\keyword{datasets}
