% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapICD.R
\name{mapICD}
\alias{mapICD}
\title{Map ICD-10 codes into the WHO 2016 cause list}
\usage{
mapICD(x)
}
\arguments{
\item{x}{a character object or a vector of ICD-10 codes}
}
\description{
Map ICD-10 codes into the WHO 2016 cause list
}
\examples{
mapICD("A90")
mapICD(c("A90", "C30"))
}
