\docType{data}
\name{RandomPhysician}
\alias{RandomPhysician}
\title{100 records of Sample Input together with two simulated physician codes}
\format{100 arbitrary input records.}
\description{
This is the same dataset as in \code{RandomVA2} with additional columns specifying physician ID and codes. 
}
\examples{
data(RandomPhysician)
head(RandomPhysician[, 1:10])
}
\keyword{datasets}
