\docType{data}
\name{RandomVA2}
\alias{RandomVA2}
\title{100 records of Sample Input}
\format{100 arbitrary input records.}
\description{
This is a  dataset consisting of 1000 arbitrary sample input 
deaths in the default format of InSilicoVA with additional columns specifying age and sex, which could be served as characteristics in sub-population estimation. 
}
\examples{
data(RandomVA2)
dim(RandomVA2)
head(RandomVA2)
}
\keyword{datasets}
