\docType{data}
\name{SampleCategory}
\alias{SampleCategory}
\title{Correspondence between InterVA causes and the physician coded cause categories}
\format{matrix of 2 columns}
\description{
This is the matrix explaining the correspondence between InterVA causes and the physician coded cause categories.
}
\examples{
data(SampleCategory)
head(SampleCategory)
}
\keyword{datasets}
