\docType{data}
\name{SampleInput_insilico}
\alias{SampleInput_insilico}
\title{100 records of Sample Input}
\format{100 arbitrary input records.}
\description{
This is a dataset consisting of 10 arbitrary sample input
deaths in the acceptable format of InSilicoVA and InterVA4. 
Any data that needs to be analyzed by this package should 
be in the same format. The orders of the input fields must 
not be changed.
}
\examples{
data(SampleInput_insilico)
}
\keyword{datasets}
