\docType{data}
\name{SamplePhysician}
\alias{SamplePhysician}
\title{100 records of Sample debiased physician codes}
\format{100 arbitrary input records.}
\description{
This is in the same format of the output running \code{\link{physician_debias}}. It is a data frame of 100 rows, and column represents ID and probability of the cause in each category.
}
\examples{
data(SamplePhysician)
head(SamplePhysician)
}
\keyword{datasets}
