\docType{data}
\name{RandomVA1}
\alias{RandomVA1}
\title{100 records of Sample Input}
\format{100 arbitrary input records.}
\description{
This is a  dataset consisting of 1000 arbitrary sample input 
deaths in the default format of InSilicoVA, i.e., the same input format as in InterVA-4 software and R package. 
}
\examples{
data(RandomVA1)
dim(RandomVA1)
head(RandomVA1)
}
\keyword{datasets}
