\docType{data}
\name{RandomVA2}
\alias{RandomVA2}
\title{100 records of Sample Input}
\format{100 arbitrary input records.}
\description{
This is a  dataset consisting of 100 arbitrary sample input 
deaths in the acceptable format of InSilicoVA. It is in the
format of a data frame. 
}
\examples{
data(RandomVA2)
dim(RandomVA2)
head(RandomVA2[, 1:10])
}
\keyword{datasets}
